Project should now rquire no configuration.
Login defaults to Karaf's in-built JAAS realm, so you no longer have to be on the University of Melbourne network to log in to Apromore.
Configuration for LDAP is now an extra step described in the README document.
Database name (apromore2) is no longer hardcoded into db-mysql.sql; it can be configured as part of the apromore-datasource-mysql feature in feature.xml.